requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dataclasses==0.6
defusedxml==0.7.1
Django==3.0.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
djongo==1.3.3
djoser==2.1.0
dnspython==2.1.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.0
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
